+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `Query::set()` and `Query::delete()` now throw  an exception if any operation fails
 - Queries attempt to detect and prevent duplicate joins
 - Model ID properties are now provided by `Model::getIDProperties()` instead of `Model::$ids`. `Model::getIDProperties()` is now a protected method instead of public.
+- Any `autoDefinition` methods should return an array of properties that will be added to the list of model properties
 
 ### Fixed
 - Rollback database transaction after uncaught exception during model persistence.